requirements.txt
langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
langchain-google-genai>=0.1.3
pydantic>=2.7
requests>=2.31
pandas>=2.1
numpy>=1.26
python-dotenv>=1.0
